+++ src/biz/bokhorst/xprivacy/PrivacyManager.java
@@ -103,6 +103,7 @@ public class PrivacyManager {
 	public final static String cSettingFirstRun = "FirstRun";
 	public final static String cSettingNotify = "Notify";
 	public final static String cSettingAndroidUsage = "AndroidUsage";
+	public final static String cSettingExtraUsage = "ExtraUsage";
 	public final static String cSettingLog = "Log";
 	public final static String cSettingExpert = "Expert";
 	public final static String cSettingRandom = "Random@boot";
@@ -373,6 +374,13 @@ public class PrivacyManager {
 			return false;
 
 		return PrivacyManager.getSettingBool(null, null, 0, PrivacyManager.cSettingAndroidUsage, false, false);
+	}
+
+	public static boolean isExtraUsageDataEnabled(int uid) {
+		if (!isUsageDataEnabled(uid))
+			return false;
+
+		return PrivacyManager.getSettingBool(null, null, 0, PrivacyManager.cSettingExtraUsage, false, false);
 	}
 
 	public static void sendUsageData(final XHook hook, Context context) {
@@ -1094,6 +1102,8 @@ public class PrivacyManager {
 			if (mName.equals(PrivacyManager.cSettingVersion))
 				return false;
 			if (mName.equals(PrivacyManager.cSettingAndroidUsage))
+				return false;
+			if (mName.equals(PrivacyManager.cSettingExtraUsage))
 				return false;
 			return (mTimestamp + cSettingCacheTimeoutMs < new Date().getTime());
 		}